feat: separate account handler
